refactor(cli): add the dispatch seam and binary
The last of task 8.8. Adapters gathers the seams a run reaches the
world through, so a whole invocation is testable against fakes; main
holds only what a test cannot have, the real streams and an exit code.

Exit codes become u8, which is what an exit code is, so main hands one
to ExitCode without a fallible conversion in the way.

Template keys are the entry types' own names rather than their
CSL-JSON strings, where article means a preprint. A configuration file
is written by a person and has to mean what a person means.

Model: claude-opus-5

+++ crates/borax/src/session.rs
@@ -28,13 +28,13 @@ pub enum Outcome {
 }
 
 /// The exit code of a run that completed with nothing skipped.
-pub const SUCCESS: i32 = 0;
+pub const SUCCESS: u8 = 0;
 
 /// The exit code of a run that never started.
 ///
 /// The code a shell treats as ordinary failure, because an invocation
 /// that could not be understood is the ordinary failure.
-pub const FATAL: i32 = 1;
+pub const FATAL: u8 = 1;
 
 /// The exit code of a completed run that skipped at least one file.
 ///
@@ -42,11 +42,11 @@ pub const FATAL: i32 = 1;
 /// with skips as a failure, while a script that wants to tell "some
 /// files need attention" from "borax could not run" reads the two codes
 /// apart.
-pub const PARTIAL: i32 = 2;
+pub const PARTIAL: u8 = 2;
 
 impl Outcome {
     /// The process exit code this outcome leaves behind.
-    pub fn code(self) -> i32 {
+    pub fn code(self) -> u8 {
         match self {
             Outcome::Success => SUCCESS,
             Outcome::Partial => PARTIAL,
